typing: finalize typing module
