+++ falcon/request.py
@@ -40,8 +40,7 @@ import warnings
 from falcon import errors
 from falcon import request_helpers as helpers
 from falcon import util
-from falcon._typing import ReadableIO
-from falcon._typing import StoreArgument
+from falcon._typing import StoreArg
 from falcon._typing import UNSET
 from falcon._typing import UnsetOr
 from falcon.constants import DEFAULT_MEDIA_TYPE
@@ -51,6 +50,7 @@ from falcon.forwarded import Forwarded
 from falcon.media import Handlers
 from falcon.media.json import _DEFAULT_JSON_HANDLER
 from falcon.stream import BoundedStream
+from falcon.typing import ReadableIO
 from falcon.util import deprecation
 from falcon.util import ETag
 from falcon.util import structures
@@ -1374,7 +1374,7 @@ class Request:
         self,
         name: str,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[str] = ...,
     ) -> str: ...
 
@@ -1383,7 +1383,7 @@ class Request:
         self,
         name: str,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: str,
     ) -> str: ...
@@ -1393,7 +1393,7 @@ class Request:
         self,
         name: str,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[str] = None,
     ) -> Optional[str]: ...
 
@@ -1401,7 +1401,7 @@ class Request:
         self,
         name: str,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[str] = None,
     ) -> Optional[str]:
         """Return the raw value of a query string parameter as a string.
@@ -1486,7 +1486,7 @@ class Request:
         required: Literal[True],
         min_value: Optional[int] = ...,
         max_value: Optional[int] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[int] = ...,
     ) -> int: ...
 
@@ -1497,7 +1497,7 @@ class Request:
         required: bool = ...,
         min_value: Optional[int] = ...,
         max_value: Optional[int] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: int,
     ) -> int: ...
@@ -1509,7 +1509,7 @@ class Request:
         required: bool = ...,
         min_value: Optional[int] = ...,
         max_value: Optional[int] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[int] = ...,
     ) -> Optional[int]: ...
 
@@ -1519,7 +1519,7 @@ class Request:
         required: bool = False,
         min_value: Optional[int] = None,
         max_value: Optional[int] = None,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[int] = None,
     ) -> Optional[int]:
         """Return the value of a query string parameter as an int.
@@ -1599,7 +1599,7 @@ class Request:
         required: Literal[True],
         min_value: Optional[float] = ...,
         max_value: Optional[float] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[float] = ...,
     ) -> float: ...
 
@@ -1610,7 +1610,7 @@ class Request:
         required: bool = ...,
         min_value: Optional[float] = ...,
         max_value: Optional[float] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: float,
     ) -> float: ...
@@ -1622,7 +1622,7 @@ class Request:
         required: bool = ...,
         min_value: Optional[float] = ...,
         max_value: Optional[float] = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[float] = ...,
     ) -> Optional[float]: ...
 
@@ -1632,7 +1632,7 @@ class Request:
         required: bool = False,
         min_value: Optional[float] = None,
         max_value: Optional[float] = None,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[float] = None,
     ) -> Optional[float]:
         """Return the value of a query string parameter as an float.
@@ -1710,7 +1710,7 @@ class Request:
         self,
         name: str,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[UUID] = ...,
     ) -> UUID: ...
 
@@ -1719,7 +1719,7 @@ class Request:
         self,
         name: str,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: UUID,
     ) -> UUID: ...
@@ -1729,7 +1729,7 @@ class Request:
         self,
         name: str,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[UUID] = ...,
     ) -> Optional[UUID]: ...
 
@@ -1737,7 +1737,7 @@ class Request:
         self,
         name: str,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[UUID] = None,
     ) -> Optional[UUID]:
         """Return the value of a query string parameter as an UUID.
@@ -1810,7 +1810,7 @@ class Request:
         self,
         name: str,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         blank_as_true: bool = ...,
         default: Optional[bool] = ...,
     ) -> bool: ...
@@ -1820,7 +1820,7 @@ class Request:
         self,
         name: str,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         blank_as_true: bool = ...,
         *,
         default: bool,
@@ -1831,7 +1831,7 @@ class Request:
         self,
         name: str,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         blank_as_true: bool = ...,
         default: Optional[bool] = ...,
     ) -> Optional[bool]: ...
@@ -1840,7 +1840,7 @@ class Request:
         self,
         name: str,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         blank_as_true: bool = True,
         default: Optional[bool] = None,
     ) -> Optional[bool]:
@@ -1924,7 +1924,7 @@ class Request:
         transform: None = ...,
         *,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[List[str]] = ...,
     ) -> List[str]: ...
 
@@ -1934,7 +1934,7 @@ class Request:
         name: str,
         transform: Callable[[str], _T],
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[List[_T]] = ...,
     ) -> List[_T]: ...
 
@@ -1944,7 +1944,7 @@ class Request:
         name: str,
         transform: None = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: List[str],
     ) -> List[str]: ...
@@ -1955,7 +1955,7 @@ class Request:
         name: str,
         transform: Callable[[str], _T],
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: List[_T],
     ) -> List[_T]: ...
@@ -1966,7 +1966,7 @@ class Request:
         name: str,
         transform: None = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[List[str]] = ...,
     ) -> Optional[List[str]]: ...
 
@@ -1976,7 +1976,7 @@ class Request:
         name: str,
         transform: Callable[[str], _T],
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[List[_T]] = ...,
     ) -> Optional[List[_T]]: ...
 
@@ -1985,7 +1985,7 @@ class Request:
         name: str,
         transform: Optional[Callable[[str], _T]] = None,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[List[_T]] = None,
     ) -> Optional[List[_T] | List[str]]:
         """Return the value of a query string parameter as a list.
@@ -2085,7 +2085,7 @@ class Request:
         format_string: str = ...,
         *,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[datetime] = ...,
     ) -> datetime: ...
 
@@ -2095,7 +2095,7 @@ class Request:
         name: str,
         format_string: str = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: datetime,
     ) -> datetime: ...
@@ -2106,7 +2106,7 @@ class Request:
         name: str,
         format_string: str = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[datetime] = ...,
     ) -> Optional[datetime]: ...
 
@@ -2115,7 +2115,7 @@ class Request:
         name: str,
         format_string: str = '%Y-%m-%dT%H:%M:%SZ',
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[datetime] = None,
     ) -> Optional[datetime]:
         """Return the value of a query string parameter as a datetime.
@@ -2169,7 +2169,7 @@ class Request:
         format_string: str = ...,
         *,
         required: Literal[True],
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[py_date] = ...,
     ) -> py_date: ...
 
@@ -2179,7 +2179,7 @@ class Request:
         name: str,
         format_string: str = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         *,
         default: py_date,
     ) -> py_date: ...
@@ -2190,7 +2190,7 @@ class Request:
         name: str,
         format_string: str = ...,
         required: bool = ...,
-        store: StoreArgument = ...,
+        store: StoreArg = ...,
         default: Optional[py_date] = ...,
     ) -> Optional[py_date]: ...
 
@@ -2199,7 +2199,7 @@ class Request:
         name: str,
         format_string: str = '%Y-%m-%d',
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[py_date] = None,
     ) -> Optional[py_date]:
         """Return the value of a query string parameter as a date.
@@ -2245,7 +2245,7 @@ class Request:
         self,
         name: str,
         required: bool = False,
-        store: StoreArgument = None,
+        store: StoreArg = None,
         default: Optional[Any] = None,
     ) -> Any:
         """Return the decoded JSON value of a query string parameter.
